Renaming queue for clarity
